requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Deep Learning
tensorflow==2.15.0
keras==2.15.0

# Image Processing
Pillow==10.1.0
numpy==1.24.3

# Data Processing (for training/analysis)
pandas==2.1.3
scikit-learn==1.3.2

# Visualization (optional, for analysis)
matplotlib==3.8.2

# Additional utilities
requests==2.31.0